==> Payroll Final Project/Windows Payroll/Payroll System/4 One Week Statistic Generator.ipynb <==
import pandas as pd
import csv
import re

user_name = input("User Name: ")
directory1 = "C:\\Users\\"
directory2 = "\\Desktop\\Payroll System"
directory = directory1 + user_name + directory2

#Looks at all Out-Day.csv's to combine into a single week statistic
folder_name = str(input("Name of Folder: "))
out_folder_name = "Out-" + folder_name

monday_file = str(input("Monday of the week: "))
tuesday_file = str(input("Tuesday of the week: "))
wednesday_file = str(input("Wednesday of the week: "))
thursday_file = str(input("Thursday of the week: "))
friday_file = str(input("Friday of the week: "))
saturday_file = str(input("Saturday of the week: "))
sunday_file = str(input("Sunday of the week: "))


out_directory_folder_monday = directory + "\\" + folder_name + "\\" + out_folder_name + "\\Out-" + monday_file+ ".csv"
out_directory_folder_tueday = directory + "\\" + folder_name + "\\" + out_folder_name + "\\Out-" + tuesday_file + ".csv"
out_directory_folder_wednesday = directory + "\\" + folder_name + "\\" + out_folder_name + "\\Out-" + wednesday_file + ".csv"
out_directory_folder_thursday = directory + "\\" + folder_name + "\\" + out_folder_name + "\\Out-" + thursday_file + ".csv"
out_directory_folder_friday = directory + "\\" + folder_name + "\\" + out_folder_name + "\\Out-" + friday_file+ ".csv"
out_directory_folder_saturday = directory + "\\" + folder_name + "\\" + out_folder_name + "\\Out-" + saturday_file + ".csv"
out_directory_folder_sunday = directory + "\\" + folder_name + "\\" + out_folder_name + "\\Out-" + sunday_file + ".csv"

Monday_df = pd.read_csv(out_directory_folder_monday)
Tuesday_df = pd.read_csv(out_directory_folder_tueday)
Wednesday_df = pd.read_csv(out_directory_folder_wednesday)
Thursday_df = pd.read_csv(out_directory_folder_thursday)
Friday_df = pd.read_csv(out_directory_folder_friday)
Saturday_df = pd.read_csv(out_directory_folder_saturday)
Sunday_df = pd.read_csv(out_directory_folder_sunday)

Employee_Info_W_Tax_df = pd.read_csv(directory + '\\1 Information CSVs\\Employee-Information-With-Tax.csv')

ID_to_Rates_df = Employee_Info_W_Tax_df.loc[:,"ID Number":"Rate Per Month"]
Tax_Info_df = Employee_Info_W_Tax_df.loc[:,"Income Tax":"SSS Employer Contribution Day"]
Tax_Info_df['ID Number'] = Employee_Info_W_Tax_df.loc[:,"ID Number"]

Business_Hours_df = pd.read_csv(directory + '\\4 Reference CSVs\\One Week Statistic Reference.csv')

Summary_Statistic_df = pd.merge(ID_to_Rates_df,Business_Hours_df,how='left')
Summary_Statistic_df = pd.merge(Summary_Statistic_df,Tax_Info_df,on='ID Number',how='left')

#Basic Business Hours, Deductibes, and Bonus
Summary_Statistic_df.loc[:,"Total Work Hours"] = Monday_df.loc[:,"Work Hours"] + Tuesday_df.loc[:,"Work Hours"] + Wednesday_df.loc[:,"Work Hours"] + Thursday_df.loc[:,"Work Hours"] + Friday_df.loc[:,"Work Hours"] + Saturday_df.loc[:,"Work Hours"] + Sunday_df.loc[:,"Work Hours"] 
Summary_Statistic_df.loc[:,"Total Basic Salary"] = Monday_df.loc[:,"Basic Salary"] + Tuesday_df.loc[:,"Basic Salary"] + Wednesday_df.loc[:,"Basic Salary"] + Thursday_df.loc[:,"Basic Salary"] + Friday_df.loc[:,"Basic Salary"] + Saturday_df.loc[:,"Basic Salary"] + Sunday_df.loc[:,"Basic Salary"] 
Summary_Statistic_df.loc[:,"Total Overtime Hours"] = Monday_df.loc[:,"Overtime Hours"] + Tuesday_df.loc[:,"Overtime Hours"] + Wednesday_df.loc[:,"Overtime Hours"] + Thursday_df.loc[:,"Overtime Hours"] + Friday_df.loc[:,"Overtime Hours"] + Saturday_df.loc[:,"Overtime Hours"] + Sunday_df.loc[:,"Overtime Hours"]
Summary_Statistic_df.loc[:,"Total Overtime Pay"] = Monday_df.loc[:,"Overtime Pay"] + Tuesday_df.loc[:,"Overtime Pay"] + Wednesday_df.loc[:,"Overtime Pay"] + Thursday_df.loc[:,"Overtime Pay"] + Friday_df.loc[:,"Overtime Pay"] + Saturday_df.loc[:,"Overtime Pay"] + Sunday_df.loc[:,"Overtime Pay"]
Summary_Statistic_df.loc[:,"Total Late Hours"] = Monday_df.loc[:,"Late Hours"] + Tuesday_df.loc[:,"Late Hours"] + Wednesday_df.loc[:,"Late Hours"] + Thursday_df.loc[:,"Late Hours"] + Friday_df.loc[:,"Late Hours"] + Saturday_df.loc[:,"Late Hours"] + Sunday_df.loc[:,"Late Hours"]
Summary_Statistic_df.loc[:,"Total Late Deduction"] = Monday_df.loc[:,"Late Deductions"] + Tuesday_df.loc[:,"Late Deductions"] + Wednesday_df.loc[:,"Late Deductions"] + Thursday_df.loc[:,"Late Deductions"] + Friday_df.loc[:,"Late Deductions"] + Saturday_df.loc[:,"Late Deductions"] + Sunday_df.loc[:,"Late Deductions"]
Summary_Statistic_df.loc[:,"Total Gross Salary"] = Monday_df.loc[:,"Gross Salary"] + Tuesday_df.loc[:,"Gross Salary"] + Wednesday_df.loc[:,"Gross Salary"] + Thursday_df.loc[:,"Gross Salary"] + Friday_df.loc[:,"Gross Salary"] + Saturday_df.loc[:,"Gross Salary"] + Sunday_df.loc[:,"Gross Salary"]
Summary_Statistic_df.loc[:,"Total Debt"] = Monday_df.loc[:,"Debt"] + Tuesday_df.loc[:,"Debt"] + Wednesday_df.loc[:,"Debt"] + Thursday_df.loc[:,"Debt"] + Friday_df.loc[:,"Debt"] + Saturday_df.loc[:,"Debt"] + Sunday_df.loc[:,"Debt"]
Summary_Statistic_df.loc[:,"Total Deductibles"] = Monday_df.loc[:,"Deductibles"] + Tuesday_df.loc[:,"Deductibles"] + Wednesday_df.loc[:,"Deductibles"] + Thursday_df.loc[:,"Deductibles"] + Friday_df.loc[:,"Deductibles"] + Saturday_df.loc[:,"Deductibles"] + Sunday_df.loc[:,"Deductibles"]
Summary_Statistic_df.loc[:,"Total Bonus"] = Monday_df.loc[:,"Bonus"] + Tuesday_df.loc[:,"Bonus"] + Wednesday_df.loc[:,"Bonus"] + Thursday_df.loc[:,"Bonus"] + Friday_df.loc[:,"Bonus"] + Saturday_df.loc[:,"Bonus"] + Sunday_df.loc[:,"Bonus"]
Summary_Statistic_df.loc[:,"Total Standing Salary"] = Monday_df.loc[:,"Standing Salary"] + Tuesday_df.loc[:,"Standing Salary"] + Wednesday_df.loc[:,"Standing Salary"] + Thursday_df.loc[:,"Standing Salary"] + Friday_df.loc[:,"Standing Salary"] + Saturday_df.loc[:,"Standing Salary"] + Sunday_df.loc[:,"Standing Salary"]

#Taxes per week
Summary_Statistic_df.loc[:,"Total Income Tax"] = Summary_Statistic_df.loc[:,"Income Tax Day"] * 7
Summary_Statistic_df.loc[:,"Total Philhealth"] = Summary_Statistic_df.loc[:,"Philhealth Day"] * 7
Summary_Statistic_df.loc[:,"Total Pagibig"] = Summary_Statistic_df.loc[:,"Pagibig Day"] * 7
Summary_Statistic_df.loc[:,"Total SSS Employee Contribution"] = Summary_Statistic_df.loc[:,"SSS Employee Contribution Day"] * 7
Summary_Statistic_df.loc[:,"Total SSS Employer Contribution"] = Summary_Statistic_df.loc[:,"SSS Employer Contribution Day"] * 7

#Final Total Salary
Summary_Statistic_df.loc[:,"Salary for the Week"] = (Summary_Statistic_df.loc[:,"Total Standing Salary"] - (Summary_Statistic_df.loc[:,"Total Income Tax"] + Summary_Statistic_df.loc[:,"Total Philhealth"] + Summary_Statistic_df.loc[:,"Total Pagibig"] + Summary_Statistic_df.loc[:,"Total SSS Employee Contribution"])).round(2)

Summary_Statistic_df

summary_statistic_name = input("Name of one week Statistic: ")
directory_folder = directory + "\\" + folder_name + "\\" + summary_statistic_name + ".csv"

Output_csv = Summary_Statistic_df.to_csv(directory_folder)
